Create api test suite
