+++ api/.gitignore
@@ -15,6 +15,7 @@ var/
 *.egg-info/
 .installed.cfg
 *.egg
+.pytest_cache/
 
 # Serverless directories
 .serverless
