+++ api/tests/__init__.py
@@ -1,0 +1,14 @@
+import json
+
+def api_gateway_event(payload: dict) -> dict:
+    return {
+        'resource': '',
+        'path': '',
+        'httpMethod': '',
+        'headers': {},
+        'requestContext': {
+            'resourcePath': '',
+            'httpMethod': ''
+        },
+        'body': json.dumps(payload)
+    }
